+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v4.12.0 (XXXX 2024)
   - Update Dradis links in README
+  - Fix the TypeError around the plugins template caching
   - Remove template_service and add mapping_service to apply mappings on tool upload
   - Include mapping module when integrations provide 'upload'